Refactor game as main class running the game
